server/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-formtools==2.5.1
django-oauth-toolkit==2.4.0
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
phonenumbers==8.13.40
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
redis==5.0.7
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
daphne>=4.0.0
dj-database-url>=0.5.0
channels>=4.0.0
channels-redis>=4.0.0
pymemcache>=3.0.0
